refactor(survey): 설문 채팅 패널 책임 분리 (#270)
